[Frontend] Relationship TreeMap widget refetch (#14448)
